--- dbms/src/Functions/FunctionsReinterpret.cpp
@@ -1,26 +1,0 @@
-#include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsReinterpret.h>
-
-
-namespace DB
-{
-
-void registerFunctionsReinterpret(FunctionFactory & factory)
-{
-    factory.registerFunction<FunctionReinterpretAsUInt8>();
-    factory.registerFunction<FunctionReinterpretAsUInt16>();
-    factory.registerFunction<FunctionReinterpretAsUInt32>();
-    factory.registerFunction<FunctionReinterpretAsUInt64>();
-    factory.registerFunction<FunctionReinterpretAsInt8>();
-    factory.registerFunction<FunctionReinterpretAsInt16>();
-    factory.registerFunction<FunctionReinterpretAsInt32>();
-    factory.registerFunction<FunctionReinterpretAsInt64>();
-    factory.registerFunction<FunctionReinterpretAsFloat32>();
-    factory.registerFunction<FunctionReinterpretAsFloat64>();
-    factory.registerFunction<FunctionReinterpretAsDate>();
-    factory.registerFunction<FunctionReinterpretAsDateTime>();
-    factory.registerFunction<FunctionReinterpretAsString>();
-    factory.registerFunction<FunctionReinterpretAsFixedString>();
-}
-
-}
